+++ SAAB_Bluetooth_AUX/Debug/objects.list
@@ -67,13 +67,13 @@
 "HAL_Driver/Src/stm32f3xx_ll_usart.o"
 "HAL_Driver/Src/stm32f3xx_ll_utils.o"
 "Utilities/STM32F3xx_Nucleo_32/stm32f3xx_nucleo_32.o"
-"src/CAN_Mailman.o"
+"src/CAN/CAN_Mailman.o"
+"src/CAN/Custom_CAN_HAL.o"
 "src/CDC_Emulator.o"
-"src/Custom_CAN_HAL.o"
-"src/DebugSerial.o"
-"src/RN52_Handler.o"
 "src/main.o"
 "src/stm32f3xx_it.o"
 "src/syscalls.o"
 "src/system_stm32f3xx.o"
+"src/Debug/DebugSerial.o"
+"src/RN52/RN52_Handler.o"
 "startup/startup_stm32f303x8.o"
